# dist/swagger.yaml
swagger: '2.0'
info:
  description: >-
    # Introduction

    Welcome to the reference for the Quire REST API!


    <a href="http://en.wikipedia.org/wiki/REST_API" target="_blank">REST</a> is
    a web-service protocol that lends itself to rapid development by using
    everyday HTTP and JSON technology.


    The Quire REST API provides a broad set of operations and resources that:

      * Consistently doing repetitive or tedious tasks.
      * Chaining a process together by responding to changes.
      * Pulling information from other locations like email or Evernote into Quire.
      * Customizing Asana for your team’s processes and workflows.

    Want to share your opinion on how our API works for you? <a
    href="https://quire.io/w/Quire_Feedbacks" target="_blank">Tell us how you
    feel </a>about using our API and what we can do to make it better.


    # API Changelog

    You can find the <a href="https://github.com/quire-api/quire-api"
    target="_blank">Changelog</a> of the API Reference in the Quire Community.


    # Authentication

    ## OAuth v2.0


    Quire recommends that you use OAuth v2.0 to authenticate to the Quire REST
    API. 


    Quire recommends you to create a dedicated API user with API write access on
    a tenant when authenticating via OAuth, and then create an OAuth client for
    this user. See <a
    href="https://knowledgecenter.zuora.com/CF_Users_and_Administrators/A_Administrator_Settings/Manage_Users/Create_an_API_User"
...


...


...

...

...

...

...
host: quire.io
basePath: /api
...
schemes:
...
paths:
...
  '/tags/{oid}':
    get:
      tags:
        - tags
      summary: Get a tag
      description: returns the full tag record.
      operationId: getTag
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of tag that needs to be fetched.
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Tag'
    put:
      tags:
        - tags
      summary: Update a tag
      description: >-
        A specific, existing tag can be updated by making a PUT request on the
        URL for that tag.

        Returns the complete updated tag record.
      operationId: updateTag
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of tag that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: User to update
          required: true
          schema:
            $ref: '#/definitions/UpdateTagBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Tag'
    delete:
      tags:
        - tags
      summary: Delete a tag
      description: >-
        A specific, existing tag can be deleted by making a DELETE request on
        the URL for that tag.
      operationId: deleteTag
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of tag that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: ok
          examples:
            application/json:
              Success: 'true'
...
  '/users/{oid}':
    get:
      tags:
        - users
      summary: Get a user
      description: returns the full user record.
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: >-
            Oid of user that needs to be fetched, can be one of an email
            address.

            Example: "john@gmail.com","me"
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
    put:
      tags:
        - users
      summary: Update a user
      description: >-
        A specific, existing user can be updated by making a PUT request on the
        URL for that user.

        Returns the complete updated user record.
      operationId: updateUser
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of user that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: User to update
          required: true
          schema:
            $ref: '#/definitions/UpdateUserBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
securityDefinitions:
  api_key:
    type: apiKey
    name: access_token
    in: header
  quire_auth:
    type: oauth2
    authorizationUrl: 'https://quire.io/oauth/authorize'
    flow: accessCode
    scopes:
      'write:projects': modify projects in your account
      'read:projects': read your projects
      basic: to read a user's profile info and media (granted by default)
      comments: to post and delete comments on a user's behalf
definitions:
...
  Tag:
    type: object
    properties:
      name:
        type: string
        example: Later
      oid:
        type: string
        example: PxXtvzXGaC0wzm7uIE92g1j1
      color:
        type: string
        example: '#a9cb77'
      createdBy:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        position: 99
      createdAt:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 99
...
  UpdateTagBody:
    type: object
    required:
      - name
    properties:
      name:
        type: string
        example: New tag
...
  UpdateUserBody:
    type: object
    properties:
      name:
        type: string
        example: John
  User:
    type: object
    properties:
      name:
        type: string
        example: John
      id:
        type: string
        example: john
      email:
        type: string
        example: john@gmail.com
      oid:
        type: string
        example: ZFrV7UGOI2oMIz48cV5IzvzZ
      image:
        type: string
        example: 'https://quire.s3.amazonaws.com/oid/image.jpg'
      initials:
        type: string
        example: J
      iconColor:
        type: string
        example: '#a9cb77'