# src/main/java/io/quire/api/model/tag/UpdateTagBody.java
package io.quire.api.model.tag;

import io.swagger.annotations.ApiModelProperty;

public class UpdateTagBody {
    @ApiModelProperty(example = "New tag", required = true)
    public String getName() { return null; }
}