# src/main/java/io/quire/api/model/user/UpdateUserBody.java
package io.quire.api.model.user;

import io.swagger.annotations.ApiModelProperty;

public class UpdateUserBody {
    @ApiModelProperty(example = "John")
    public String getName() { return null; }
}